refactor: split host DNS handler
Split handlers into separate files to make maintaining them easier.
